cleaned up

+++ Point.cpp
@@ -1,10 +1,8 @@
 #include "Point.hpp"
 
-Point::Point(int L, double l)
+Point::Point(int L, int l)
     :longitude(L), latitude(l) {}
 
 Point::getdistance(Point p1, Point p2) {
-
-
     return 0;
 }
